pyproject.toml
[project]
name = "kleinanzeigen-monitor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]